피드백 반영 수정

+++ src/main/java/com/example/bootproject/post/controller/PostController.java
@@ -1,5 +1,6 @@
 package com.example.bootproject.post.controller;
 
+import com.example.bootproject.enums.DeleteType;
 import com.example.bootproject.post.domain.Post;
 import com.example.bootproject.post.domain.PostSearch;
 import com.example.bootproject.post.dto.PostDetailDTO;
@@ -7,6 +8,7 @@ import com.example.bootproject.post.dto.PostRegistDTO;
 import com.example.bootproject.post.dto.PostUpdateDTO;
 import com.example.bootproject.post.service.PostService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -36,27 +38,31 @@ public class PostController {
 
     @GetMapping("/api/posts")
     public List<PostDetailDTO> searchPost(PostSearch search) {
-        List<PostDetailDTO> postDetailDTOS = postService.postListSearch(search);
+        List<PostDetailDTO> postDetails = postService.postListSearch(search);
 
-        return postDetailDTOS;
+        return postDetails;
     }
 
     @PostMapping("/api/posts")
     public ResponseEntity<PostDetailDTO> insertPost(@RequestBody PostRegistDTO post) {
-        PostDetailDTO postDetailDTO = postService.insertPost(post);
+        PostDetailDTO postDetail = postService.savePost(post);
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(postDetailDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body(postDetail);
     }
 
     @PutMapping("/api/posts/{postId}")
     public PostUpdateDTO updatePost(@PathVariable("postId") Long id, @RequestBody PostRegistDTO postRegistDTO) {
-        PostUpdateDTO postUpdateDTO = postService.updatePost(postRegistDTO, id);
+        PostUpdateDTO postUpdate = postService.updatePost(postRegistDTO, id);
 
-        return postUpdateDTO;
+        return postUpdate;
     }
 
     @DeleteMapping("/api/posts/{postId}")
-    public void deletePost(@PathVariable("postId") Long id) {
-        postService.deletePost(id);
+    public void deletePost(@PathVariable("postId") Long id, @RequestParam("deleteType") DeleteType deleteType) {
+        if (deleteType.equals(DeleteType.HARD)){
+            postService.deletePost(id);
+        } else {
+            postService.softDeletePost(id);
+        }
     }
 }
